0.7 Audio changes
Added:
AudioControllerExtender, that is in every scene and is used to controll settings sliders and buttons to play audios. That is required because AudioController is notDestoryedOnLoad object.

Instead of using PlayerPrefs, game now uses new settings class and JSON to save player audio sliders settings.
Audio will be set to 0.5f if the game is opened for the first time.